release: v4.2.4 - compilation fix, SMP/TLB hardening, and syscall quality
## Compilation Blockers (3)
- BUG-PG-SUBTREE: Implement free_pagetable_subtree() (pagetable.c)
- BUG-SELECT-FILES: Fix current->files to fd_get() (syscall.c)
- BUG-SOCKETPAIR-TCP: Replace TCP with pipe() for socketpair (syscall.c)

## Memory Management (5)
- BUG-MEM-SLAB: Fix kfree slab page reclaim address (mem.c)
- BUG-PML4-LEAK: Free PML4[0] deep-copied subtree on exit (pagetable.c)
- BUG-REF-UNDERFLOW: Handle ref_count underflow in free_pagetable (pagetable.c)
- BUG-BUDDY-REF: Reset ref_count in buddy_split (mem.c)
- BUG-ALLOC-RETRY: Retry buddy_split on failure (mem.c)

## SMP/TLB (5)
- BUG-SPLIT-TLB: Use smp_tlb_shootdown in split_huge_page (pagetable.c)
- BUG-HUGE-TLB: Use smp_tlb_shootdown in map_huge_page_2mb (pagetable.c)
- BUG-COW-TLB: Use smp_tlb_shootdown in COW path (pagetable.c)
- BUG-MAP-UAF: Free old page after new PTE write (pagetable.c)
- BUG-HUGE-TOCTOU: Atomic ref_count in map_huge_page_2mb (pagetable.c)

## Security (4)
- BUG-READV-TOCTOU: Validate each iov_base in readv (syscall.c)
- BUG-WRITEV-TOCTOU: Validate each iov_base in writev (syscall.c)
- BUG-IRELATIVE-SMEP: Disable SMEP for IRELATIVE resolver (elfloader.c)
- BUG-IOPL: Mask IOPL bits in sigreturn RFLAGS (signal.c)

## Syscall Fixes (5)
- BUG-GETDENTS-OFFSET: Update file offset after getdents64 (syscall.c)
- BUG-SELECT-TIMEOUT: Implement select timeout (syscall.c)
- BUG-SETSOCKOPT-VALIDATE: Validate sockfd in setsockopt (syscall.c)
- BUG-ARGV-LIMIT: Dynamic argv scanning in execve (syscall.c)
- BUG-POLL-WRITABLE: Check writability for POLLOUT (syscall.c)

## Scheduling (3)
- BUG-FIND-UAF: Increment ref_count in find_task_by_pid (sched.c)
- BUG-WAITPID-ERRNO: Set ECHILD for non-child waitpid (sched.c)
- BUG-SMP-ENQUEUE: Add lock to smp_enqueue_task (sched.c)

## Code Quality (3)
- BUG-SIG-UB: Use uint64_t for pending/blocked in signal.h
- BUG-1GB-HUGE: Handle 1GB huge pages in clone_current_pml4 (pagetable.c)
- BUG-FREE-ORDER: Document free_page order=0 (mem.c)

## Documentation
- Version: v4.2.3 -> v4.2.4 (version.h, README.md, CHANGELOG.md)
- syscall count: 77 -> 85 (api.md)
- module_sign: Build-time key injection mechanism (module_sign.c)

+++ kernel/signal.h
@@ -58,8 +58,13 @@ struct sigframe {
 
 struct signal_state {
     struct sigaction actions[NSIG];
-    uint32_t         pending;
-    uint32_t         blocked;
+    /* FIXED (v4.2.4): Use uint64_t for pending/blocked bitmasks.
+     * With NSIG=32, bit 31 (1<<31) is within uint32_t range, but
+     * the expression (1U << 31) is undefined behavior in C (shifting
+     * into sign bit).  Using uint64_t avoids this UB and allows
+     * future expansion beyond 32 signals.  (BUG-SIG-UB) */
+    uint64_t         pending;
+    uint64_t         blocked;
     uint64_t         saved_rsp;  /* for sigreturn: original user RSP */
     uint64_t         saved_rip;  /* for sigreturn: original user RIP */
 };
